// _OLD/src/interfaces/comic-series.interface.ts
/**
 * Interface representing a comic series in the database.
 * 
 * @interface ComicSeries
 * @property {number} id - The unique identifier for the comic series.
 * @property {string} name - The name of the comic series.
 * @property {string} folder_path - The path to the folder containing the comic series.
 * @property {string} description - A brief description of the comic series.
 * @property {string} publisher - The publisher of the comic series.
 * @property {number} start_year - The year the comic series started.
 * @property {number | null} end_year - The year the comic series ended, or null if it is ongoing.
 * @property {number} total_issues - The total number of issues in the comic series.
 */
export interface ComicSeries {
    id: number | null; // This will be set after inserting the series
    name: string;
    folder_path: string;
    description: string | null;
    publisher: string | null;
    start_year: number | null;
    end_year: number | null;
    total_issues: number | null;
    library_id: number; // The ID of the library this series belongs to
}

/**
 * Type for inserting a new comic series into the database.
 * This type omits the 'id' property, as it will be generated by the database.
 * @type ComicSeriesInsert
 * @property {string} name - The name of the comic series.
 * @property {string} folder_path - The path to the folder containing the comic series.
 * @property {string | null} description - A brief description of the comic series.
 * @property {string | null} publisher - The publisher of the comic series.
 * @property {number | null} start_year - The year the comic series started.
 * @property {number | null} end_year - The year the comic series ended, or null if it is ongoing.
 * @property {number | null} total_issues - The total number of issues in the comic series.
 */
export type ComicSeriesInsert = Omit<ComicSeries, 'id'> & { id?: number | null };